+++ app/Http/Controllers/Recipe/RecipeLikeController.php
@@ -5,9 +5,9 @@ namespace App\Http\Controllers\Recipe;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Recipe\ShortVersion\ShortRecipeCollection;
 use App\Models\Recipe;
-use App\Services\Like\Interfaces\LikeCreator;
-use App\Services\Like\Interfaces\LikeDeleter;
-use App\Services\Like\Interfaces\LikeGetter;
+use App\Services\Like\Contracts\LikeCreator;
+use App\Services\Like\Contracts\LikeDeleter;
+use App\Services\Like\Contracts\LikeGetter;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\ResourceCollection;
